AI_Code_Review_Phase1_Updated/requirements.txt
Django
djangorestframework
psycopg2
django-allauth
django-rest-auth
django-cors-headers
openai